checkin-service/requirements.txt
Flask==3.0.3
pymongo==4.7.2
python-dotenv==1.0.1
pytest==8.2.2
pytest-cov==5.0.0